move Pet's services to another package add ComponentScan annotation
