+++ src/main/java/com/springframework/pets/CatPetService.java
@@ -1,4 +1,4 @@
-package guru.springframework.sfgdi.services;
+package com.springframework.pets;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
