+++ src/main/java/com/springframework/pets/PetService.java
@@ -1,4 +1,4 @@
-package guru.springframework.sfgdi.services;
+package com.springframework.pets;
 
 /**
  * Created by jt on 12/28/19.
