+++ src/main/java/guru/springframework/sfgdi/SfgDiApplication.java
@@ -4,7 +4,9 @@ import guru.springframework.sfgdi.controllers.*;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.ComponentScan;
 
+@ComponentScan(basePackages = {"guru.springframework.sfgdi","com.springframework.pets"})
 @SpringBootApplication
 public class SfgDiApplication {
